+++ Final Project/Source/views/header.html
@@ -8,7 +8,7 @@
                 <li><a href="/index" class="active">Home</a></li>
                 <li><a href="/menu">Menu</a></li>
                 <li><a href="/reservation">Reservation</a></li>
-                <li><a href="/order">Order Online</a></li>
+                <li><a href="/order">Your Order</a></li>
             </ul>
         </div>
     </div>
