+++ Final Project/source/views/order.html
@@ -1,0 +1,160 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Order - Vietnamese Restaurant</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+    <link rel="stylesheet" href="./css/order.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
+</head>
+
+<body>
+    <h1>Your Order</h1>
+
+    <!-- Display Order Items -->
+    <form action="/order/complete" method="post">
+        <table class="table">
+            <thead>
+                <tr>
+                    <th>Dish</th>
+                    <th>Name</th>
+                    <th>Quantity</th>
+                    <th>Price</th>
+                    <th>Total</th>
+                    <th>Actions</th>
+                </tr>
+            </thead>
+            <tbody>
+                <% let total=0; %>
+                    <% orderItems.forEach((item, index)=> { %>
+                        <tr id="row-<%= index %>">
+                            <!-- ... (rest of the row structure) ... -->
+                            <!-- Hidden input fields for each row -->
+                            <input type="hidden" name="items[<%= index %>][dish_id]" value="<%= item.dish.id %>">
+                            <input type="hidden" name="items[<%= index %>][name]" value="<%= item.dish.name %>">
+                            <input type="hidden" name="items[<%= index %>][quantity]" id="input-quantity-<%= index %>"
+                                value="<%= item.quantity %>">
+                            <input type="hidden" name="items[<%= index %>][price]" value="<%= item.dish.price %>">
+                            <% total +=item.dish.price * item.quantity; %>
+                                <td><img src='<%= item.dish.image_url %>' class="dish-image"></td>
+                                <td>
+                                    <%= item.dish.name %>
+                                </td>
+                                <td>
+                                    <button onclick="decreaseQuantity(<%= index %>);return false;">-</button>
+                                    <span id="quantity-<%= index %>">
+                                        <%= item.quantity %>
+                                    </span>
+                                    <button onclick="increaseQuantity(<%= index %>);return false;">+</button>
+                                </td>
+                                <td id="price-<%= index %>">
+                                    <%= item.dish.price %>
+                                </td>
+                                <td id="total-<%= index %>">
+                                    <%= item.dish.price * item.quantity %>
+                                </td>
+                                <td>
+                                    <a onclick="deleteItem(<%= index %>)"><span class="bi bi-trash"></span></a>
+                                </td>
+                        </tr>
+                        <% }); %>
+                            <tr>
+                                <th></th>
+                                <th></th>
+                                <th></th>
+                                <th></th>
+                                <th></th>
+                                <th><b>Total Price: <span id="grandTotal">
+                                            <%= total %>
+                                        </span></b></th>
+                            </tr>
+            </tbody>
+        </table>
+        <!-- Shipment Information -->
+        <div id="shipment-div-container">
+            <div id="shipment-div">
+                <h2>Shipment Information</h2>
+                <label for="name">Name:</label>
+                <input type="text" id="name" name="name" required>
+
+                <label for="address">Address:</label>
+                <input type="text" id="address" name="address" required>
+
+                <label for="phone">Phone:</label>
+                <input type="tel" id="phone" name="phone" required>
+                <input type="hidden" name="orderId" value="<%= id %>"> <!-- Assuming you have the order ID available -->
+                <input type="hidden" name="totalPrice" value="<%= total %>">
+                <!-- Assuming you have the order ID available -->
+
+                <label for="email">Email:</label>
+                <input type="email" id="email" name="email" required> <!-- Email input for sending confirmation -->
+                <div>
+                    <input type="submit" value="Complete Order">
+                </div>
+            </div>
+        </div>
+    </form>
+
+    <script>
+        function increaseQuantity(index) {
+            const quantityElem = document.getElementById(`quantity-${index}`);
+            const priceElem = document.getElementById(`price-${index}`);
+            const totalElem = document.getElementById(`total-${index}`);
+            const grandTotalElem = document.getElementById('grandTotal');
+
+            let quantity = parseInt(quantityElem.innerText);
+            let price = parseFloat(priceElem.innerText);
+
+            quantity += 1;
+            quantityElem.innerText = quantity;
+            totalElem.innerText = (price * quantity).toFixed(2);
+
+            const quantityInput = document.getElementById(`input-quantity-${index}`);
+            quantityInput.value = quantity;
+
+            updateGrandTotal();
+        }
+
+        function decreaseQuantity(index) {
+            const quantityElem = document.getElementById(`quantity-${index}`);
+            const priceElem = document.getElementById(`price-${index}`);
+            const totalElem = document.getElementById(`total-${index}`);
+            const grandTotalElem = document.getElementById('grandTotal');
+
+            let quantity = parseInt(quantityElem.innerText);
+            let price = parseFloat(priceElem.innerText);
+
+            if (quantity > 1) {
+                quantity -= 1;
+                quantityElem.innerText = quantity;
+                totalElem.innerText = (price * quantity).toFixed(2);
+                const quantityInput = document.getElementById(`input-quantity-${index}`);
+                quantityInput.value = quantity;
+            }
+            updateGrandTotal();
+        }
+
+        function updateGrandTotal() {
+            const grandTotalElem = document.getElementById('grandTotal');
+            let grandTotal = 0;
+
+            document.querySelectorAll('[id^="total-"]').forEach(totalElem => {
+                grandTotal += parseFloat(totalElem.innerText);
+            });
+
+            grandTotalElem.innerText = grandTotal.toFixed(2);
+        }
+
+        function deleteItem(index) {
+            // Remove the table row
+            const row = document.getElementById(`row-${index}`);
+            row.parentNode.removeChild(row);
+
+            // Update the hidden input fields
+            updateGrandTotal();
+        }
+
+    </script>
+</body>
+
+</html>
